moved update_cache to views folder
